meson: Reorganize toolchain wrappers
Prior to this change toolchain wrapper script required additional
arguments to find flag files and to select the compiler.

Such approach has issues when extending it with the rust support.

From now toolchain wrapper and flags are located in the same directory.
Compiler is selected using symlinks in a similar war busybox / toybox
works.

Here's how the wrapper directory looks after this change:

    $ ls -l toolchain_wrapper/

    -rw-rw-r-- exec.cc
    -rw-rw-r-- exec.cflags
    -rw-rw-r-- exec.cppflags
    -rw-rw-r-- exec.cxx
    -rw-rw-r-- exec.link_args
    -rw-rw-r-- sharedlib.cc
    -rw-rw-r-- sharedlib.cflags
    -rw-rw-r-- sharedlib.cppflags
    -rw-rw-r-- sharedlib.cxx
    -rw-rw-r-- sharedlib.link_args
    lrwxrwxrwx wrap_c -> ./wrapper.sh
    lrwxrwxrwx wrap_cxx -> ./wrapper.sh
    -rwxrwxr-x wrapper.sh
    lrwxrwxrwx wrap_rust_ld -> ./wrapper.sh

Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>